+++ scripts/smoke-test-webgl.py
@@ -52,10 +52,19 @@ class RequestVerifier:
         else:
             raise Exception(info)
 
-    def ExpectMessage(self, index, substring):
+    def CheckMessage(self, index, substring, negate):
         message = self.__requests[index]["body"]
-        self.Expect("HTTP Request #{} contains \"{}\".".format(index, substring),
-                    substring in message or substring.replace("'", "\"") in message)
+        contains = substring in message or substring.replace(
+            "'", "\"") in message
+        return contains if not negate else not contains
+
+    def ExpectMessage(self, index, substring):
+        self.Expect("HTTP Request #{} contains \"{}\".".format(
+            index, substring), self.CheckMessage(index, substring, False))
+
+    def ExpectMessageNot(self, index, substring):
+        self.Expect("HTTP Request #{} doesn't contain \"{}\".".format(
+            index, substring), self.CheckMessage(index, substring, True))
 
 
 t = RequestVerifier()
@@ -68,11 +77,12 @@ class Handler(SimpleHTTPRequestHandler):
     def do_POST(self):
         body = ""
         content = self.rfile.read(int(self.headers['Content-Length']))
-        try:
-            body = content.decode("utf-8")
-        except:
-            logging.exception("Exception while parsing an API request")
-            body = binascii.hexlify(bytearray(content))
+        parts = content.split(b'\n')
+        for part in parts:
+            try:
+                body += '\n' + part.decode("utf-8")
+            except:
+                body += '\n(binary chunk: {} bytes)'.format(len(part))
         t.Capture(self.requestline, body)
         self.send_response(HTTPStatus.OK, '{'+'}')
         self.end_headers()
@@ -135,9 +145,15 @@ t.ExpectMessage(currentMessage, "'type':'session'")
 currentMessage += 1
 t.ExpectMessage(currentMessage, "'type':'event'")
 t.ExpectMessage(currentMessage, "LogError(GUID)")
+# t.ExpectMessage(
+#     currentMessage, "'filename':'screenshot.jpg','attachment_type':'event.attachment'")
+# t.ExpectMessageNot(currentMessage, "'length':0")
 currentMessage += 1
 t.ExpectMessage(currentMessage, "'type':'event'")
 t.ExpectMessage(currentMessage, "CaptureMessage(GUID)")
+# t.ExpectMessage(
+#     currentMessage, "'filename':'screenshot.jpg','attachment_type':'event.attachment'")
+# t.ExpectMessageNot(currentMessage, "'length':0")
 currentMessage += 1
 t.ExpectMessage(currentMessage, "'type':'event'")
 t.ExpectMessage(
@@ -147,4 +163,7 @@ t.ExpectMessage(currentMessage, "'extra':{'extra-key':42}")
 t.ExpectMessage(currentMessage, "'tags':{'tag-key':'tag-value'")
 t.ExpectMessage(
     currentMessage, "'user':{'email':'email@example.com','id':'user-id','ip_address':'::1','username':'username','other':{'role':'admin'}}")
+# t.ExpectMessage(
+# currentMessage, "'filename':'screenshot.jpg','attachment_type':'event.attachment'")
+# t.ExpectMessageNot(currentMessage, "'length':0")
 print('TEST: PASS', flush=True)