+++ src/Sentry.Unity.Editor/ScriptableSentryUnityOptionsEditor.cs
@@ -29,6 +29,10 @@ namespace Sentry.Unity.Editor
             EditorGUILayout.TextField("Release Override", options.ReleaseOverride);
             EditorGUILayout.TextField("Environment Override", options.EnvironmentOverride);
             EditorGUILayout.Toggle("Attach Stacktrace", options.AttachStacktrace);
+            EditorGUILayout.Toggle("Attach Screenshot", options.AttachScreenshot);
+            EditorGUILayout.IntField("Screenshot Max Height", options.ScreenshotMaxHeight);
+            EditorGUILayout.IntField("Screenshot Max Width", options.ScreenshotMaxWidth);
+            EditorGUILayout.IntField("Screenshot Quality", options.ScreenshotQuality);
             EditorGUILayout.IntField("Max Breadcrumbs", options.MaxBreadcrumbs);
             EditorGUILayout.EnumPopup("Report Assemblies Mode", options.ReportAssembliesMode);
             EditorGUILayout.Toggle("Send Default Pii", options.SendDefaultPii);