+++ src/Sentry.Unity/SentryOptionsUtility.cs
@@ -19,6 +19,10 @@ namespace Sentry.Unity
             scriptableOptions.AutoSessionTracking = options.AutoSessionTracking;
 
             scriptableOptions.AttachStacktrace = options.AttachStacktrace;
+            scriptableOptions.AttachScreenshot = options.AttachScreenshot;
+            scriptableOptions.ScreenshotMaxWidth = options.ScreenshotMaxWidth;
+            scriptableOptions.ScreenshotMaxHeight = options.ScreenshotMaxHeight;
+            scriptableOptions.ScreenshotQuality = options.ScreenshotQuality;
             scriptableOptions.MaxBreadcrumbs = options.MaxBreadcrumbs;
             scriptableOptions.ReportAssembliesMode = options.ReportAssembliesMode;
             scriptableOptions.SendDefaultPii = options.SendDefaultPii;