+++ test/Sentry.Unity.Tests/ScriptableSentryUnityOptions.cs
@@ -78,6 +78,10 @@ namespace Sentry.Unity.Tests
                 AutoSessionTracking = false,
                 AutoSessionTrackingInterval = TimeSpan.FromSeconds(1),
                 AttachStacktrace = true,
+                AttachScreenshot = true,
+                ScreenshotMaxWidth = 1,
+                ScreenshotMaxHeight = 1,
+                ScreenshotQuality = 1,
                 MaxBreadcrumbs = 1,
                 ReportAssembliesMode = ReportAssembliesMode.None,
                 SendDefaultPii = true,
@@ -104,6 +108,10 @@ namespace Sentry.Unity.Tests
             scriptableOptions.AutoSessionTracking = expectedOptions.AutoSessionTracking;
             scriptableOptions.AutoSessionTrackingInterval = (int)expectedOptions.AutoSessionTrackingInterval.TotalMilliseconds;
             scriptableOptions.AttachStacktrace = expectedOptions.AttachStacktrace;
+            scriptableOptions.AttachScreenshot = expectedOptions.AttachScreenshot;
+            scriptableOptions.ScreenshotMaxWidth = expectedOptions.ScreenshotMaxWidth;
+            scriptableOptions.ScreenshotMaxHeight = expectedOptions.ScreenshotMaxHeight;
+            scriptableOptions.ScreenshotQuality = expectedOptions.ScreenshotQuality;
             scriptableOptions.MaxBreadcrumbs = expectedOptions.MaxBreadcrumbs;
             scriptableOptions.ReportAssembliesMode = expectedOptions.ReportAssembliesMode;
             scriptableOptions.SendDefaultPii = expectedOptions.SendDefaultPii;
@@ -166,6 +174,10 @@ namespace Sentry.Unity.Tests
             Assert.AreEqual(expected.AutoSessionTracking, actual.AutoSessionTracking);
             Assert.AreEqual(expected.AutoSessionTrackingInterval, actual.AutoSessionTrackingInterval);
             Assert.AreEqual(expected.AttachStacktrace, actual.AttachStacktrace);
+            Assert.AreEqual(expected.AttachScreenshot, actual.AttachScreenshot);
+            Assert.AreEqual(expected.ScreenshotMaxWidth, actual.ScreenshotMaxWidth);
+            Assert.AreEqual(expected.ScreenshotMaxHeight, actual.ScreenshotMaxHeight);
+            Assert.AreEqual(expected.ScreenshotQuality, actual.ScreenshotQuality);
             Assert.AreEqual(expected.MaxBreadcrumbs, actual.MaxBreadcrumbs);
             Assert.AreEqual(expected.ReportAssembliesMode, actual.ReportAssembliesMode);
             Assert.AreEqual(expected.SendDefaultPii, actual.SendDefaultPii);